Refactoring / updates for PR comments
to finish this PR off.
