+++ lib/api/WebApi.js
@@ -10,7 +10,6 @@ import CompositeId from '@/lib/io/CompositeId';
 import AuthState from '@/lib/model/AuthState';
 import Category from '@/lib/model/Category';
 import CollisionEvent from '@/lib/model/CollisionEvent';
-import JobMetadata from '@/lib/model/JobMetadata';
 import Joi from '@/lib/model/Joi';
 import Study from '@/lib/model/Study';
 import StudyRequest from '@/lib/model/StudyRequest';
@@ -19,6 +18,10 @@ import StudyRequestComment from '@/lib/model/StudyRequestComment';
 import SuccessResponse from '@/lib/model/SuccessResponse';
 import User from '@/lib/model/User';
 import CentrelineLocation from '@/lib/model/helpers/CentrelineLocation';
+import {
+  normalizeJobMetadata,
+  normalizeJobMetadatas,
+} from '@/lib/model/helpers/NormalizeUtils';
 import DateTime from '@/lib/time/DateTime';
 
 const apiClient = new AxiosBackendClient('/api');
@@ -84,13 +87,12 @@ async function getCollisionsByCentrelineTotal(features) {
 
 async function getJob(jobId) {
   const jobMetadata = await schedulerClient.fetch(`/jobs/${jobId}`);
-  return JobMetadata.read.validateAsync(jobMetadata);
+  return normalizeJobMetadata(jobMetadata);
 }
 
 async function getJobs() {
   const jobMetadatas = await schedulerClient.fetch('/jobs');
-  const jobMetadatasSchema = Joi.array().items(JobMetadata.read);
-  return jobMetadatasSchema.validateAsync(jobMetadatas);
+  return normalizeJobMetadatas(jobMetadatas);
 }
 
 async function getLocationsByCentreline(features) {
@@ -387,7 +389,7 @@ async function postJobGenerateCollisionReports(csrf, features, filters, reportFo
     data,
   };
   const jobMetadata = await schedulerClient.fetch('/jobs/generateCollisionReports', options);
-  return JobMetadata.read.validateAsync(jobMetadata);
+  return normalizeJobMetadata(jobMetadata);
 }
 
 async function postJobGenerateStudyReports(csrf, features, filters, reportFormat) {
@@ -403,7 +405,7 @@ async function postJobGenerateStudyReports(csrf, features, filters, reportFormat
     data,
   };
   const jobMetadata = await schedulerClient.fetch('/jobs/generateStudyReports', options);
-  return JobMetadata.read.validateAsync(jobMetadata);
+  return normalizeJobMetadata(jobMetadata);
 }
 
 async function postStudyRequest(csrf, studyRequest) {
@@ -440,7 +442,7 @@ async function putJobCancel(csrf, jobMetadata) {
     csrf,
   };
   const persistedJobMetadata = await schedulerClient.fetch(url, options);
-  return JobMetadata.read.validateAsync(persistedJobMetadata);
+  return normalizeJobMetadata(persistedJobMetadata);
 }
 
 async function putJobDismiss(csrf, jobMetadata) {
@@ -451,7 +453,7 @@ async function putJobDismiss(csrf, jobMetadata) {
     csrf,
   };
   const persistedJobMetadata = await schedulerClient.fetch(url, options);
-  return JobMetadata.read.validateAsync(persistedJobMetadata);
+  return normalizeJobMetadata(persistedJobMetadata);
 }
 
 async function putStudyRequest(csrf, studyRequest) {
