+++ lib/db/JobMetadataDAO.js
@@ -1,30 +1,10 @@
 import db from '@/lib/db/db';
 import JobDescription from '@/lib/jobs/JobDescription';
 import JobType from '@/lib/jobs/JobType';
-import JobMetadata from '@/lib/model/JobMetadata';
-import Joi from '@/lib/model/Joi';
-
-async function normalizeJobMetadata(jobMetadata) {
-  const jobMetadataNormalized = await JobMetadata.read.validateAsync(jobMetadata);
-  const { metadata, type } = jobMetadataNormalized;
-  const metadataNormalized = await type.validateMetadata(metadata);
-  jobMetadataNormalized.metadata = metadataNormalized;
-  return jobMetadataNormalized;
-}
-
-async function normalizeJobMetadatas(jobMetadatas) {
-  const jobMetadatasSchema = Joi.array().items(JobMetadata.read);
-  const jobMetadatasNormalized = await jobMetadatasSchema.validateAsync(jobMetadatas);
-  const tasks = jobMetadatasNormalized.map(
-    ({ metadata, type }) => type.validateMetadata(metadata),
-  );
-  const metadatasNormalized = await Promise.all(tasks);
-  const n = jobMetadatasNormalized.length;
-  for (let i = 0; i < n; i++) {
-    jobMetadatasNormalized[i].metadata = metadatasNormalized[i];
-  }
-  return jobMetadatasNormalized;
-}
+import {
+  normalizeJobMetadata,
+  normalizeJobMetadatas,
+} from '@/lib/model/helpers/NormalizeUtils';
 
 /**
  * Data access layer for background job metadata.
