+++ lib/io/storage/StoragePath.js
@@ -35,7 +35,7 @@ class StoragePath {
     } else {
       const hashBuilder = crypto.createHash('md5');
       hashBuilder.update(JSON.stringify(options));
-      const hash = hashBuilder.digest('hex').slice(0, 8);
+      const hash = hashBuilder.digest('hex');
       options = `_${hash}`;
     }
