+++ lib/jobs/JobPoller.js
@@ -1,4 +1,5 @@
 import { getJob } from '@/lib/api/WebApi';
+import DateTime from '@/lib/time/DateTime';
 import DateTimeZone from '@/lib/time/DateTimeZone';
 
 const INTERVAL_TICK = 1000;
@@ -73,31 +74,9 @@ class JobPoller extends EventTarget {
       return null;
     }
 
-    let s = Math.floor(this.timeRemaining / 1000);
-    let m = Math.floor(s / 60);
-    const h = Math.floor(m / 60);
-    m %= 60;
-    s %= 60;
-
-    if (h > 1) {
-      return 'a while';
-    }
-    if (h === 1) {
-      return `1 hour ${m} minutes`;
-    }
-    if (m > 1) {
-      return `${m} minutes`;
-    }
-    if (m === 1) {
-      return `1 minute ${s} seconds`;
-    }
-    if (s > 1) {
-      return `${s} seconds`;
-    }
-    if (s === 1) {
-      return '1 second';
-    }
-    return 'now';
+    return DateTime.local()
+      .plus(this.timeRemaining)
+      .toRelative({ locale: 'en-US' });
   }
 
   async updateGetJob() {
@@ -136,12 +115,20 @@ class JobPoller extends EventTarget {
     }
 
     const now = DateTimeZone.utc();
-    let elapsed = now.valueOf() - startedAt.valueOf();
-    if (elapsed < INTERVAL_GET_JOB) {
-      elapsed = INTERVAL_GET_JOB;
+    const elapsed = now.valueOf() - startedAt.valueOf();
+    if (elapsed <= 0) {
+      /*
+       * Since `startedAt` is measured server-side and `now` is measured client-side,
+       * clock drift can cause `now` to be before `startedAt`!  In this case, it's
+       * impossible to estimate remaining time without knowing how much clock drift
+       * there is, so we don't bother.
+       */
+      this.timeRemaining = null;
+      return;
     }
+    const elapsedAdjusted = elapsed + INTERVAL_GET_JOB;
     const f = progressCurrent / progressTotal;
-    const timeRemaining = Math.round(elapsed * (1 - f) / f);
+    const timeRemaining = Math.round(elapsedAdjusted * (1 - f) / f);
 
     this.timeRemaining = timeRemaining;
     this.progressCurrentPrev = progressCurrent;
