+++ lib/jobs/JobRunnerGenerateReports.js
@@ -88,8 +88,6 @@ class JobRunnerGenerateReports extends JobRunnerBase {
       const report = reports[i];
       /* eslint-disable-next-line no-await-in-loop */
       const savedNew = await this.saveReport(report);
-      /* eslint-disable-next-line no-await-in-loop */
-      await this.incrProgressCurrent();
 
       if (savedNew) {
         // Random delay strategy; see `DELAY_MIN` / `DELAY_MAX` above.
@@ -97,6 +95,14 @@ class JobRunnerGenerateReports extends JobRunnerBase {
         /* eslint-disable-next-line no-await-in-loop */
         await asyncDelay(wait);
       }
+
+      /*
+       * We increment progress after the random delay so that `JobPoller` is less likely to
+       * underestimate remaining time.  (From a user psychology perspective, it's better to
+       * overestimate than underestimate here!)
+       */
+      /* eslint-disable-next-line no-await-in-loop */
+      await this.incrProgressCurrent();
     }
     return this.zipReports(reports);
   }
