+++ lib/model/helpers/NormalizeUtils.js
@@ -1,4 +1,6 @@
 import { CentrelineType } from '@/lib/Constants';
+import JobMetadata from '@/lib/model/JobMetadata';
+import Joi from '@/lib/model/Joi';
 
 function intersectionToFeature({
   geom,
@@ -17,6 +19,28 @@ function intersectionToFeature({
     lat,
     lng,
   };
+}
+
+async function normalizeJobMetadata(jobMetadata) {
+  const jobMetadataNormalized = await JobMetadata.read.validateAsync(jobMetadata);
+  const { metadata, type } = jobMetadataNormalized;
+  const metadataNormalized = await type.validateMetadata(metadata);
+  jobMetadataNormalized.metadata = metadataNormalized;
+  return jobMetadataNormalized;
+}
+
+async function normalizeJobMetadatas(jobMetadatas) {
+  const jobMetadatasSchema = Joi.array().items(JobMetadata.read);
+  const jobMetadatasNormalized = await jobMetadatasSchema.validateAsync(jobMetadatas);
+  const tasks = jobMetadatasNormalized.map(
+    ({ metadata, type }) => type.validateMetadata(metadata),
+  );
+  const metadatasNormalized = await Promise.all(tasks);
+  const n = jobMetadatasNormalized.length;
+  for (let i = 0; i < n; i++) {
+    jobMetadatasNormalized[i].metadata = metadatasNormalized[i];
+  }
+  return jobMetadatasNormalized;
 }
 
 function segmentToFeature({
@@ -56,11 +80,15 @@ function segmentToFeature({
 
 const NormalizeUtils = {
   intersectionToFeature,
+  normalizeJobMetadata,
+  normalizeJobMetadatas,
   segmentToFeature,
 };
 
 export {
   NormalizeUtils as default,
   intersectionToFeature,
+  normalizeJobMetadata,
+  normalizeJobMetadatas,
   segmentToFeature,
 };
